# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Cервер на Golang с OpenAPI документацией.",
        "title": "Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Maksim",
            "url": "https://github.com/hhullen",
            "email": "hhullen@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/product": {
            "get": {
                "description": "Возвращает продукт.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Возвращает продукт",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"c85a189d-d173-42e2-8e00-54395234d93d\"",
                        "description": "uid",
                        "name": "uid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "true",
                        "description": "avoid_cache",
                        "name": "avoid_cache",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавление продукта. Если продукт существует, обновится цена, количество в наличии, дата обновления и вернется uid этого продукта.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Добавление продукта",
                "parameters": [
                    {
                        "description": "Информация о продукте",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.AddProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаление продукта.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Удаление продукта",
                "parameters": [
                    {
                        "description": "uid",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DeleteProductResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            },
            "patch": {
                "description": "Убавление количества продукта.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Убавление количества продукта",
                "parameters": [
                    {
                        "description": "uid и уколичество",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/datastruct.DecreaseProductsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DecreaseProductsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.DecreaseProductsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Возвращает список продуктов. Если offset limit равны 0, вернет список всех",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Возвращает список продуктов",
                "parameters": [
                    {
                        "type": "string",
                        "example": "0",
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "10",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "true",
                        "description": "avoid_cache",
                        "name": "avoid_cache",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/datastruct.GetProductsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/datastruct.Status"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "datastruct.AddProductRequest": {
            "type": "object",
            "required": [
                "available_stock",
                "category",
                "image_id",
                "name",
                "price",
                "supplier_id"
            ],
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "example": 1023
                },
                "avoid_cache": {
                    "type": "boolean",
                    "example": true
                },
                "category": {
                    "type": "string",
                    "example": "construction"
                },
                "image_id": {
                    "type": "string",
                    "example": "376de312-5bcb-4320-8ba3-bd2050548229"
                },
                "last_update_date": {
                    "type": "string",
                    "example": "31.01.2026"
                },
                "name": {
                    "type": "string",
                    "example": "Wooden beam"
                },
                "price": {
                    "type": "number",
                    "example": 299.95
                },
                "supplier_id": {
                    "type": "string",
                    "example": "609ccf6f-7fb4-44bd-aa77-bc9e0e7572b4"
                },
                "uid": {
                    "type": "string",
                    "example": "c85a189d-d173-42e2-8e00-54395234d93d"
                }
            }
        },
        "datastruct.AddProductResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
...
        "datastruct.DecreaseProductsRequest": {
            "type": "object",
            "required": [
                "amount",
                "uid"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 3
                },
                "uid": {
                    "type": "string",
                    "example": "c85a189d-d173-42e2-8e00-54395234d93d"
                }
            }
        },
        "datastruct.DecreaseProductsResponse": {
            "type": "object",
            "properties": {
                "left": {
                    "type": "integer"
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
...
        "datastruct.DeleteProductRequest": {
            "type": "object",
            "required": [
                "uid"
            ],
            "properties": {
                "avoid_cache": {
                    "type": "boolean",
                    "example": true
                },
                "uid": {
                    "type": "string",
                    "example": "c85a189d-d173-42e2-8e00-54395234d93d"
                }
            }
        },
        "datastruct.DeleteProductResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
...
        "datastruct.GetProductResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "product": {
                    "$ref": "#/definitions/datastruct.Product"
                },
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
        "datastruct.GetProductsResponse": {
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/datastruct.Product"
                    }
                }
            }
        },
...
        "datastruct.Product": {
            "type": "object",
            "required": [
                "available_stock",
                "category",
                "image_id",
                "name",
                "price",
                "supplier_id"
            ],
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "example": 1023
                },
                "category": {
                    "type": "string",
                    "example": "construction"
                },
                "image_id": {
                    "type": "string",
                    "example": "376de312-5bcb-4320-8ba3-bd2050548229"
                },
                "last_update_date": {
                    "type": "string",
                    "example": "31.01.2026"
                },
                "name": {
                    "type": "string",
                    "example": "Wooden beam"
                },
                "price": {
                    "type": "number",
                    "example": 299.95
                },
                "supplier_id": {
                    "type": "string",
                    "example": "609ccf6f-7fb4-44bd-aa77-bc9e0e7572b4"
                },
                "uid": {
                    "type": "string",
                    "example": "c85a189d-d173-42e2-8e00-54395234d93d"
                }
            }
        },
        "datastruct.Status": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "status message"
                }
            }
        },
...
    }
}